+++ src/igl/CMakeLists.txt
@@ -29,6 +29,8 @@ endif()
 igl_set_cxxstd(IGLLibrary 17)
 igl_set_folder(IGLLibrary "IGL")
 
+add_subdirectory(glslang)
+
 if(IGL_WITH_OPENGL OR IGL_WITH_OPENGLES OR IGL_WITH_WEBGL)
   add_subdirectory(opengl)
   target_link_libraries(IGLLibrary PUBLIC IGLOpenGL)
