+++ src/igl/vulkan/CMakeLists.txt
@@ -20,30 +20,10 @@ endif()
 
 add_library(IGLVulkan ${SRC_FILES} ${HEADER_FILES})
 
-target_link_libraries(IGLVulkan PRIVATE IGLLibrary)
+target_link_libraries(IGLVulkan PRIVATE IGLLibrary IGLGlslang)
 
 igl_set_cxxstd(IGLVulkan 17)
 igl_set_folder(IGLVulkan "IGL")
-
-# glslang
-# cmake-format: off
-set(ENABLE_GLSLANG_BINARIES OFF CACHE BOOL "")
-set(ENABLE_HLSL             OFF CACHE BOOL "")
-set(ENABLE_CTEST            OFF CACHE BOOL "")
-set(ENABLE_OPT              OFF CACHE BOOL "")
-set(ENABLE_SPVREMAPPER      OFF CACHE BOOL "")
-set(SKIP_GLSLANG_INSTALL    ON  CACHE BOOL "")
-add_subdirectory(${IGL_ROOT_DIR}/third-party/deps/src/glslang "glslang")
-igl_set_folder(GenericCodeGen     "third-party/glslang")
-igl_set_folder(glslang            "third-party/glslang")
-igl_set_folder(MachineIndependent "third-party/glslang")
-igl_set_folder(OGLCompiler        "third-party/glslang")
-igl_set_folder(OSDependent        "third-party/glslang")
-igl_set_folder(SPIRV              "third-party/glslang")
-igl_set_folder(glslang-default-resource-limits "third-party/glslang")
-# cmake-format: on
-
-target_link_libraries(IGLVulkan PUBLIC glslang SPIRV glslang-default-resource-limits)
 
 find_package(Vulkan REQUIRED)
 
@@ -54,6 +34,7 @@ target_include_directories(IGLVulkan PUBLIC "${IGL_ROOT_DIR}/third-party/deps/sr
 target_include_directories(IGLVulkan PUBLIC "${IGL_ROOT_DIR}/third-party/deps/src/volk")
 target_include_directories(IGLVulkan PUBLIC "${IGL_ROOT_DIR}/third-party/deps/src/vma/include")
 target_include_directories(IGLVulkan PUBLIC "${IGL_ROOT_DIR}/third-party/deps/src/SPIRV-Headers/include")
+target_include_directories(IGLVulkan PUBLIC "${IGL_ROOT_DIR}/third-party/deps/src/glslang")
 
 if(WIN32)
   add_definitions("-DVK_USE_PLATFORM_WIN32_KHR=1")
