+++ src/igl/vulkan/VulkanHelpers.c
@@ -1260,48 +1260,6 @@ VkRect2D ivkGetRect2D(int32_t x, int32_t y, uint32_t width, uint32_t height) {
   return rect;
 }
 
-static glslang_stage_t getGLSLangShaderStage(VkShaderStageFlagBits stage) {
-  switch (stage) {
-  case VK_SHADER_STAGE_VERTEX_BIT:
-    return GLSLANG_STAGE_VERTEX;
-  case VK_SHADER_STAGE_TESSELLATION_CONTROL_BIT:
-    return GLSLANG_STAGE_TESSCONTROL;
-  case VK_SHADER_STAGE_TESSELLATION_EVALUATION_BIT:
-    return GLSLANG_STAGE_TESSEVALUATION;
-  case VK_SHADER_STAGE_GEOMETRY_BIT:
-    return GLSLANG_STAGE_GEOMETRY;
-  case VK_SHADER_STAGE_FRAGMENT_BIT:
-    return GLSLANG_STAGE_FRAGMENT;
-  case VK_SHADER_STAGE_COMPUTE_BIT:
-    return GLSLANG_STAGE_COMPUTE;
-  default:
-    assert(false);
-  };
-  assert(false);
-  return GLSLANG_STAGE_COUNT;
-}
-
-glslang_input_t ivkGetGLSLangInput(VkShaderStageFlagBits stage,
-                                   const glslang_resource_t* resource,
-                                   const char* shaderCode) {
-  const glslang_input_t input = {
-      .language = GLSLANG_SOURCE_GLSL,
-      .stage = getGLSLangShaderStage(stage),
-      .client = GLSLANG_CLIENT_VULKAN,
-      .client_version = GLSLANG_TARGET_VULKAN_1_1,
-      .target_language = GLSLANG_TARGET_SPV,
-      .target_language_version = GLSLANG_TARGET_SPV_1_3,
-      .code = shaderCode,
-      .default_version = 100,
-      .default_profile = GLSLANG_NO_PROFILE,
-      .force_default_version_and_profile = false,
-      .forward_compatible = false,
-      .messages = GLSLANG_MSG_DEFAULT_BIT,
-      .resource = resource,
-  };
-  return input;
-}
-
 VkResult ivkCreateShaderModuleFromSPIRV(const struct VulkanFunctionTable* vt,
                                         VkDevice device,
                                         const void* dataSPIRV,
@@ -1689,121 +1647,34 @@ VkResult ivkVmaCreateAllocator(const struct VulkanFunctionTable* vt,
   return vmaCreateAllocator(&ci, outVma);
 }
 
-void ivkGlslangResource(glslang_resource_t* glslangResource,
-                        const VkPhysicalDeviceProperties* deviceProperties) {
-  const VkPhysicalDeviceLimits* limits = deviceProperties ? &deviceProperties->limits : NULL;
+void ivkUpdateGlslangResource(glslang_resource_t* res, const VkPhysicalDeviceProperties* props) {
+  const VkPhysicalDeviceLimits* limits = props ? &props->limits : NULL;
+  if (!limits) {
+    return;
+  }
 
-  const glslang_resource_t resource = {
-      .max_lights = 32,
-      .max_clip_planes = 6,
-      .max_texture_units = 32,
-      .max_texture_coords = 32,
-      .max_vertex_attribs = limits ? (int)limits->maxVertexInputAttributes : 64,
-      .max_vertex_uniform_components = 4096,
-      .max_varying_floats = 64,
-      .max_vertex_texture_image_units = 32,
-      .max_combined_texture_image_units = 80,
-      .max_texture_image_units = 32,
-      .max_fragment_uniform_components = 4096,
-      .max_draw_buffers = 32,
-      .max_vertex_uniform_vectors = 128,
-      .max_varying_vectors = 8,
-      .max_fragment_uniform_vectors = 16,
-      .max_vertex_output_vectors = 16,
-      .max_fragment_input_vectors = 15,
-      .min_program_texel_offset = -8,
-      .max_program_texel_offset = 7,
-      .max_clip_distances = limits ? (int)limits->maxClipDistances : 8,
-      .max_compute_work_group_count_x = limits ? (int)limits->maxComputeWorkGroupCount[0] : 65535,
-      .max_compute_work_group_count_y = limits ? (int)limits->maxComputeWorkGroupCount[1] : 65535,
-      .max_compute_work_group_count_z = limits ? (int)limits->maxComputeWorkGroupCount[2] : 65535,
-      .max_compute_work_group_size_x = limits ? (int)limits->maxComputeWorkGroupSize[0] : 1024,
-      .max_compute_work_group_size_y = limits ? (int)limits->maxComputeWorkGroupSize[1] : 1024,
-      .max_compute_work_group_size_z = limits ? (int)limits->maxComputeWorkGroupSize[2] : 64,
-      .max_compute_uniform_components = 1024,
-      .max_compute_texture_image_units = 16,
-      .max_compute_image_uniforms = 8,
-      .max_compute_atomic_counters = 8,
-      .max_compute_atomic_counter_buffers = 1,
-      .max_varying_components = 60,
-      .max_vertex_output_components = limits ? (int)limits->maxVertexOutputComponents : 64,
-      .max_geometry_input_components = limits ? (int)limits->maxGeometryInputComponents : 64,
-      .max_geometry_output_components = limits ? (int)limits->maxGeometryOutputComponents : 128,
-      .max_fragment_input_components = limits ? (int)limits->maxFragmentInputComponents : 128,
-      .max_image_units = 8,
-      .max_combined_image_units_and_fragment_outputs = 8,
-      .max_combined_shader_output_resources = 8,
-      .max_image_samples = 0,
-      .max_vertex_image_uniforms = 0,
-      .max_tess_control_image_uniforms = 0,
-      .max_tess_evaluation_image_uniforms = 0,
-      .max_geometry_image_uniforms = 0,
-      .max_fragment_image_uniforms = 8,
-      .max_combined_image_uniforms = 8,
-      .max_geometry_texture_image_units = 16,
-      .max_geometry_output_vertices = limits ? (int)limits->maxGeometryOutputVertices : 256,
-      .max_geometry_total_output_components = limits ? (int)limits->maxGeometryTotalOutputComponents
-                                                     : 1024,
-      .max_geometry_uniform_components = 1024,
-      .max_geometry_varying_components = 64,
-      .max_tess_control_input_components =
-          limits ? (int)limits->maxTessellationControlPerVertexInputComponents : 128,
-      .max_tess_control_output_components =
-          limits ? (int)limits->maxTessellationControlPerVertexOutputComponents : 128,
-      .max_tess_control_texture_image_units = 16,
-      .max_tess_control_uniform_components = 1024,
-      .max_tess_control_total_output_components = 4096,
-      .max_tess_evaluation_input_components =
-          limits ? (int)limits->maxTessellationEvaluationInputComponents : 128,
-      .max_tess_evaluation_output_components =
-          limits ? (int)limits->maxTessellationEvaluationOutputComponents : 128,
-      .max_tess_evaluation_texture_image_units = 16,
-      .max_tess_evaluation_uniform_components = 1024,
-      .max_tess_patch_components = 120,
-      .max_patch_vertices = 32,
-      .max_tess_gen_level = 64,
-      .max_viewports = limits ? (int)limits->maxViewports : 16,
-      .max_vertex_atomic_counters = 0,
-      .max_tess_control_atomic_counters = 0,
-      .max_tess_evaluation_atomic_counters = 0,
-      .max_geometry_atomic_counters = 0,
-      .max_fragment_atomic_counters = 8,
-      .max_combined_atomic_counters = 8,
-      .max_atomic_counter_bindings = 1,
-      .max_vertex_atomic_counter_buffers = 0,
-      .max_tess_control_atomic_counter_buffers = 0,
-      .max_tess_evaluation_atomic_counter_buffers = 0,
-      .max_geometry_atomic_counter_buffers = 0,
-      .max_fragment_atomic_counter_buffers = 1,
-      .max_combined_atomic_counter_buffers = 1,
-      .max_atomic_counter_buffer_size = 16384,
-      .max_transform_feedback_buffers = 4,
-      .max_transform_feedback_interleaved_components = 64,
-      .max_cull_distances = limits ? (int)limits->maxCullDistances : 8,
-      .max_combined_clip_and_cull_distances = limits ? (int)limits->maxCombinedClipAndCullDistances
-                                                     : 8,
-      .max_samples = 4,
-      .max_mesh_output_vertices_nv = 256,
-      .max_mesh_output_primitives_nv = 512,
-      .max_mesh_work_group_size_x_nv = 32,
-      .max_mesh_work_group_size_y_nv = 1,
-      .max_mesh_work_group_size_z_nv = 1,
-      .max_task_work_group_size_x_nv = 32,
-      .max_task_work_group_size_y_nv = 1,
-      .max_task_work_group_size_z_nv = 1,
-      .max_mesh_view_count_nv = 4,
-      .maxDualSourceDrawBuffersEXT = 1,
-      .limits = {
-          .non_inductive_for_loops = true,
-          .while_loops = true,
-          .do_while_loops = true,
-          .general_uniform_indexing = true,
-          .general_attribute_matrix_vector_indexing = true,
-          .general_varying_indexing = true,
-          .general_sampler_indexing = true,
-          .general_variable_indexing = true,
-          .general_constant_matrix_vector_indexing = true,
-      }};
-
-  *glslangResource = resource;
+  res->max_vertex_attribs = (int)limits->maxVertexInputAttributes;
+  res->max_clip_distances = (int)limits->maxClipDistances;
+  res->max_compute_work_group_count_x = (int)limits->maxComputeWorkGroupCount[0];
+  res->max_compute_work_group_count_y = (int)limits->maxComputeWorkGroupCount[1];
+  res->max_compute_work_group_count_z = (int)limits->maxComputeWorkGroupCount[2];
+  res->max_compute_work_group_size_x = (int)limits->maxComputeWorkGroupSize[0];
+  res->max_compute_work_group_size_y = (int)limits->maxComputeWorkGroupSize[1];
+  res->max_compute_work_group_size_z = (int)limits->maxComputeWorkGroupSize[2];
+  res->max_vertex_output_components = (int)limits->maxVertexOutputComponents;
+  res->max_geometry_input_components = (int)limits->maxGeometryInputComponents;
+  res->max_geometry_output_components = (int)limits->maxGeometryOutputComponents;
+  res->max_fragment_input_components = (int)limits->maxFragmentInputComponents;
+  res->max_geometry_output_vertices = (int)limits->maxGeometryOutputVertices;
+  res->max_geometry_total_output_components = (int)limits->maxGeometryTotalOutputComponents;
+  res->max_tess_control_input_components =
+      (int)limits->maxTessellationControlPerVertexInputComponents;
+  res->max_tess_control_output_components =
+      (int)limits->maxTessellationControlPerVertexOutputComponents;
+  res->max_tess_evaluation_input_components = (int)limits->maxTessellationEvaluationInputComponents;
+  res->max_tess_evaluation_output_components =
+      (int)limits->maxTessellationEvaluationOutputComponents;
+  res->max_viewports = (int)limits->maxViewports;
+  res->max_cull_distances = (int)limits->maxCullDistances;
+  res->max_combined_clip_and_cull_distances = (int)limits->maxCombinedClipAndCullDistances;
 }
